move plotting to seperate file
